Revert "Alignment fix"
This reverts commit 97ad353cab16293807f70cf807c5ecd5c6dbfb9b.

+++ Runtime/Scripts/VehicleComponents/ROS/Core/TimeStamp.cs
@@ -32,12 +32,12 @@ namespace Unity.Robotics.Core
         // NOTE: We could define these operators in a transport-specific extension package
         public static implicit operator TimeMsg(TimeStamp stamp)
         {
-            return new TimeMsg((uint)stamp.Seconds, stamp.NanoSeconds);
+            return new TimeMsg(stamp.Seconds, stamp.NanoSeconds);
         }
 
         public static implicit operator TimeStamp(TimeMsg stamp)
         {
-            return new TimeStamp((int)stamp.sec, stamp.nanosec);
+            return new TimeStamp(stamp.sec, stamp.nanosec);
         }
     }
 }
